requirements.txt
google-api-python-client==2.77.0
google-auth-oauthlib==0.7.0
google-auth-httplib2==0.1.1
requests==2.28.1
beautifulsoup4==4.12.0
aiohttp==3.8.1
python-dotenv==0.19.2